+++ src/rest-api.js
@@ -66,7 +66,6 @@ function format(config, url) {
 }
 
 module.exports = function restAPI(config, url, method, params) {
-  if (method === "del") { method = "delete"; }
   const token = config.get("sudo") ? config.get("secret") : (config.get("accessToken") || config.get("secret"));
   const conf = {
     token,